use long long at output to fix arithmetic overflow in "112PD1/hw2/pB.c"

+++ C/112PD1/hw2/pB.c
@@ -3,6 +3,6 @@
 int main(){
 	int n,k1,k2,k3;
 	scanf("%d %d %d %d",&n,&k1,&k2,&k3);
-	printf("%d\n",n*(k1+k2+k3-(k1/3)-(k2/3)-(k3/3)));
+	printf("%lld\n",(long long)n*(k1+k2+k3-(k1/3)-(k2/3)-(k3/3)));
 	return 0;
 }
